+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Packages: Dependencies updates ([#2977](https://github.com/dso-toolkit/dso-toolkit/issues/2977))
 * Dropdown Menu: Flaky tests fix ([#3081](https://github.com/dso-toolkit/dso-toolkit/issues/3081))
 * Action List: flaky screenshot test ([#2894](https://github.com/dso-toolkit/dso-toolkit/issues/2894))
+* Document Component: flaky e2e tests verbeteren ([#2934](https://github.com/dso-toolkit/dso-toolkit/issues/2934))
 
 ## 🪼 Release 69.2.1 - 2025-03-21